you see the edit buget you cant really see the icon fix it feel like the icon are blank for all of it also make it the date formating as may 2025  also i want the app start from june 2025 not may please

+++ src/components/budget/MonthNavigator.tsx
@@ -9,7 +9,7 @@ export function MonthNavigator() {
   const { currentDisplayMonthId, navigateToPreviousMonth, navigateToNextMonth } = useBudget();
 
   const displayDate = parseYearMonth(currentDisplayMonthId);
-  const monthName = displayDate.toLocaleString('default', { month: 'long' });
+  const monthName = displayDate.toLocaleString('default', { month: 'short' }); // Changed from 'long' to 'short'
   const year = displayDate.getFullYear();
 
   return (
